start writing some testing functions
Right now we're spinning up a whole new tower-web Service for each test.
The IpAddr calls are to make sure we pick an unused port so that tests can happen in parallel.

This might not be the right way to do things since this means that we're going to spin up a whole new Service for each test (and then take it down). Also problematic is that we don't have a good way to stop the service. We can run in non-blocking mode - the ServiceBuilder will give us a Future that we can throw at a Reactor - run our tests and then stop the Reactor we're using, but that still seems a little clunky.